Reverting some of the real-number values here.

+++ remglk.h
@@ -36,8 +36,8 @@
 /* Some useful type declarations. */
 
 typedef struct grect_struct {
-    double left, top;
-    double right, bottom;
+    int left, top;
+    int right, bottom;
 } grect_t;
 
 #define grect_set_from_size(boxref, wid, hgt)   \
